Move util.js
